Translation for Calendar and Tags in Food Diary, removed unused dependencies and files
