+++ package.json
@@ -19,12 +19,8 @@
     "react-native-action-button": "^2.8.5",
     "react-native-calendar-picker": "^6.0.0",
     "react-native-dialog": "^5.6.0",
-    "react-native-dynamic-search-bar": "^0.1.11",
     "react-native-elements": "^1.1.0",
     "react-native-modal-datetime-picker": "^7.4.0",
-    "react-native-paper": "^2.15.2",
-    "react-native-search-bar": "^3.4.2",
-    "react-native-vector-icons": "^6.4.2",
     "react-navigation": "^3.0.9"
   },
   "devDependencies": {
