+++ src/assets/languages/developer.json
@@ -57,6 +57,34 @@
 		"SYMPTOM_LIST": "Zuastandlista",
 		"SEARCH": "Suacha",
 		"NICKNAME": "Spitznama",
-		"ADD_NEW_SYMPTOM": "Neues Symptom"
+		"ADD_NEW_SYMPTOM": "Neues Symptom",
+		"GLUTEN": "Gluten",
+		"NO_GLUTEN": "Kein Gluten",
+		"UNSURE": "Unsicher",
+		"BREAKFAST": "Frühstück",
+		"LUNCH": "Mittagessen",
+		"DINNER": "Abendessen",
+		"SNACK": "Snack",
+		"JAN": "Januar",
+		"FEB": "Februar",
+		"MAR": "März",
+		"APR": "April",
+		"MAY": "Mai",
+		"JUN": "Juni",
+		"JUL": "July",
+		"AUG": "August",
+		"SEP": "September",
+		"OCT": "Oktober",
+		"NOV": "Novemeber",
+		"DEC": "Dezember",
+		"PREVIOUS": "Vorheriger",
+		"NEXT": "Nächster",
+		"MON": "Mo",
+		"TUE": "Di",
+		"WED": "Mi",
+		"THU": "Do",
+		"FRI": "Fr",
+		"SAT": "Sa",
+		"SUN": "So"
 	}
 }
