+++ src/assets/languages/english.json
@@ -57,6 +57,34 @@
 		"SYMPTOM_LIST": "Symptomlist",
 		"SEARCH": "Search",
 		"NICKNAME": "Nickname",
-		"ADD_NEW_SYMPTOM": "New symptom"
+		"ADD_NEW_SYMPTOM": "New symptom",
+		"GLUTEN": "gluten",
+		"NO_GLUTEN": "No gluten",
+		"UNSURE": "unsure",
+		"BREAKFAST": "breakfast",
+		"LUNCH": "lunch",
+		"DINNER": "dinner",
+		"SNACK": "snack",
+		"JAN": "January",
+		"FEB": "February",
+		"MAR": "March",
+		"APR": "April",
+		"MAY": "May",
+		"JUN": "June",
+		"JUL": "July",
+		"AUG": "August",
+		"SEP": "September",
+		"OCT": "October",
+		"NOV": "Novemeber",
+		"DEC": "Decemeber",
+		"PREVIOUS": "Previous",
+		"NEXT": "Next",
+		"MON": "Mon",
+		"TUE": "Tue",
+		"WED": "Wed",
+		"THU": "Thu",
+		"FRI": "Fri",
+		"SAT": "Sat",
+		"SUN": "Sun"
 	}
 }
